[Bronze I] Title: 평균, Time: 104 ms, Memory: 14480 KB -BaekjoonHub

+++ 백준/Bronze/1546. 평균/평균.java
@@ -22,7 +22,7 @@ public class Main {
         StringTokenizer st = new StringTokenizer(br.readLine(), " ");
         
         for(int i = 0; i < N; i++) {
-        	arr[i] = Float.parseFloat(st.nextToken());		
+        	arr[i] = Double.parseDouble(st.nextToken());		
 
         }
